refactor: move place selector from events-app to components-package
HH-447.
