+++ packages/components/src/components/domain/place/index.ts
@@ -1,0 +1,3 @@
+export { default as PlaceText } from './PlaceText';
+export { default as useFetchPlacesByIds } from './useFetchPlacesByIds';
+export { default as PlaceSelector } from './placeSelector/PlaceSelector';
